编译错误

+++ AESever/common/timer_wheel_manager.h
@@ -66,13 +66,13 @@ struct TimerObj
     ///> 下一次触发时间，仅仅作为准确性参考值。不作为排序依据。
     std::chrono::steady_clock::time_point debug_next_triggle_time{ std::chrono::steady_clock::now() };
     ///> 单位毫秒。冷却间隔，根据cd来做排序参考值。
-    uint32_t set_cd_ms{0};
+    uint64_t set_cd_ms{0};
     ///> 需要触发次数  -1 表示无数次 
-    int32_t count{ 0 };
+    uint64_t count{ 0 };
     ///> 从计时起总的，用来打日志
-    int32_t total_cd{ 0 };
+    uint64_t total_cd{ 0 };
     ///> 下次cd
-    int32_t next_cd{ 0 };
+    uint64_t next_cd{ 0 };
     ///> 回调
     TimerCB callback{nullptr};
